update /play, remove unused styles
