+++ public/unity/Viewer/index.html
@@ -68,7 +68,7 @@
         codeUrl: buildUrl + "/Build.wasm",
         streamingAssetsUrl: "StreamingAssets",
         companyName: "Asset Layer",
-        productName: "Rolltopia",
+        productName: "Unity Model Viewer",
         productVersion: "0.1",
         showBanner: unityShowBanner,
       };
