+++ public/unity/index.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="utf-8">
     <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
-    <title>Unity WebGL Player | magic-test</title>
+    <title>Unity WebGL Player | Sample App</title>
     <link rel="shortcut icon" href="TemplateData/favicon.ico">
     <link rel="stylesheet" href="TemplateData/style.css">
   </head>
@@ -71,7 +71,7 @@
         codeUrl: buildUrl + "/WebGL.wasm",
         streamingAssetsUrl: "StreamingAssets",
         companyName: "Asset Layer",
-        productName: "Rolltopia",
+        productName: "Sample App",
         productVersion: "0.1",
         showBanner: unityShowBanner,
       };
